pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "langchain-movie-recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.5",
    "langchain-anthropic>=1.0.3",
    "python-dotenv>=1.2.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/movie_recommender"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.4",
    "ty>=0.0.1a26",
]
